Split XmlHttp into Controller/Provider
